safari-600.7-branch/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/securitypolicy-alloweval-expected.txt

PASS eval is allowed when no policy exists.
PASS eval is not allowed when policy exists.
